libraries/_to_delete/gcp-pipeline-builder/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-pipeline-builder"
version = "1.0.0"
description = "A production-grade Python framework for building GCP data pipelines"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Data Platform Team", email = "data-platform@company.com"}
]
keywords = ["gcp", "dataflow", "bigquery", "apache-beam", "airflow", "data-pipeline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "apache-beam[gcp]>=2.52.0",
    "apache-airflow>=2.5.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-pubsub>=2.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "faker>=18.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "apache-airflow-providers-google>=10.0.0",
    "apache-airflow-providers-apache-beam>=5.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-exporter-gcp-trace>=1.5.0",
]
dynatrace = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

